fix: move AppList and AppSetDefaultLinkDialog to features package
- Move AppSetDefaultLinkDialog from apps/web to packages/features/apps/components
- Create platform-compatible AppList in packages/features/apps/components
- Update web app to re-export AppSetDefaultLinkDialog from features
- Update web app imports to use features path
